apache-kafka/requirements.txt
# Kafka client library for Python
kafka-python==2.0.2

